Fix task state transitions for unique tasks.
This fixes a race condition where a unique task could be queued, moved by the
first worker to the active queue, queued again, moved by the second worker to
the active queue, and then removed by the first worker before it finished
processing on the second worker.

The fix for this problem is not trivial and is implemented by 1) disallowing
multiple unique tasks being processed at the same time (a unique task is
requeued if another instance is being processed), 2) making sure all task
transitions behave accordingly if the task is in the destination state.

As part of this, improvements were made to Redis scripts which includes adding
unit tests and more documentation.