Task refactor and task canceling
Refactored the code to make use of the Task class, which now can be used to
queue new tasks as well as cancel/retry/inspect them, and is used internally
as well.

Added a TaskNotFound exception which is raised when a task cannot be found
or does not exist in the expected queue, so we know when e.g. canceling a
scheduled task failed.