Allow custom context managers to see task errors (#165)
* tests: add failing context-manager test

if the task fails, we should get a traceback in our context managers

* worker: avoid swallowing context manager exceptions

we use user-specified context managers to wrap `self._execute_forked`
but that function always catches all exceptions. this means that our
`__exit__` function always thinks the tasks raised no error, and so do
not necesserily perform proper cleanup.

here, we move the context-manager wrapping closer to where we invoke the
function. the downside is that, if we swallow exceptions in the context
manager, we won't get retry logic, which i think is an acceptable
tradeoff

Original Author: Igor Serebryany <igor47@moomers.org>