If possible, retry tasks that fail with "execution not found" (#323)
- This will retry on "execution not found" if the task also retries `JobTimeoutError`, i.e. is idempotent.
- The "execution not found" error will be changed to a warning.

The only possible danger from this is that retries on this error would not count towards the retry counter since we don't have an execution, which might lead to too many retries at the first retry interval without a cap. I see this possibility as rare, and if we wanted to we could manually create a dummy execution.